Mergani fileovi za mobile/desktop
