+++ tv_shows_app/src/components/features/auth/LoginForm/LoginForm.tsx
@@ -12,7 +12,6 @@ import {
   Flex,
   FormControl,
   Heading,
-  Hide,
   Input,
   InputGroup,
   InputLeftElement,
@@ -24,21 +23,106 @@ import {
 import { log } from 'console';
 import NextLink from 'next/link';
 import { useRouter } from 'next/navigation';
-import { Fragment, useState } from 'react';
+import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import useSWRMutation from 'swr/mutation';
-import { LoginFormMobile } from './layouts/LoginForm.mobile';
-import { LoginFormDesktop } from './layouts/LoginForm.desktop';
+
+interface ILoginForm {
+  email: string;
+  password: string;
+}
 
 export const LoginForm = () => {
+  const router = useRouter();
+  const {
+    register,
+    handleSubmit,
+    formState: { isSubmitting },
+  } = useForm<ILoginForm>();
+  const { trigger } = useSWRMutation(swrKeys.login, mutator<ILoginForm>, {
+    onSuccess: () => {
+      router.push('/shows');
+    },
+  });
+  const onLogin = async (data: ILoginForm) => {
+    await trigger(data);
+  };
+
   return (
     <>
-      <Hide above="1024px">
-        <LoginFormMobile />
-      </Hide>
-      <Show above="1024px">
-        <LoginFormDesktop />
-      </Show>
+      <Flex
+        direction="column"
+        alignItems="center"
+        justifyContent="center"
+        gap="4"
+        backgroundColor={{ base: 'purple2', lg: 'darkPurple' }}
+        height="100vh"
+      >
+        <Flex
+          width={{ base: '288px', lg: '500px' }}
+          height={{ base: '100%', lg: '500px' }}
+          backgroundColor={{ base: 'transparent', lg: 'purple2' }}
+          direction="column"
+          alignItems="center"
+          as="form"
+          gap="30px"
+          py="10"
+          borderRadius="10"
+          onSubmit={handleSubmit(onLogin)}
+          boxShadow={{ base: 'none', lg: 'customShadow' }}
+        >
+          <Heading
+            textStyle="headline"
+            size="md"
+            marginBottom="8"
+            color="white"
+          >
+            TV shows APP
+          </Heading>
+          <InputGroup>
+            <InputLeftElement
+              pointerEvents="none"
+              children={<EmailIcon color="white" width="24px" />}
+              fontSize="md"
+              ml={{ base: 0, lg: '56px' }}
+            />
+            <FormControl textColor="white">
+              <Input
+                isDisabled={isSubmitting}
+                {...register('email')}
+                placeholder="Email"
+                type="email"
+                color="white"
+                _placeholder={{ color: 'white' }}
+                size={{ base: 'sm', lg: 'md' }}
+                data-testid="email"
+                mx={{ base: 'auto', lg: '56px' }}
+              />
+            </FormControl>
+          </InputGroup>
+          <PasswordInput
+            disable={isSubmitting}
+            thatPart={register('password')}
+            placeholder="Password"
+            testId="password"
+          />
+          <Button
+            isLoading={isSubmitting}
+            loadingText="Logging in"
+            type="submit"
+            variant="default"
+            textStyle="button"
+          >
+            LOG IN
+          </Button>
+          <Flex flexDirection="row" color="white" gap="1" fontSize="xs">
+            <Text textStyle="smallCaption"> Don't have an account? </Text>
+            <Text as={NextLink} href="/register" textStyle="smallCaptionBold">
+              Register
+            </Text>
+          </Flex>
+        </Flex>
+      </Flex>
     </>
   );
 };
