+++ tv_shows_app/src/components/features/reviews/ReviewItem/ReviewItem.tsx
@@ -10,8 +10,6 @@ import {
   MenuButton,
   MenuItem,
   MenuList,
-  Show,
-  Hide,
 } from '@chakra-ui/react';
 import { StarReview } from '../StarReview';
 import { ChevronDownIcon, DeleteIcon, SearchIcon } from '@chakra-ui/icons';
@@ -21,10 +19,6 @@ import { swrKeys } from '@/fetchers/swrKeys';
 import { deleteReviewMutator } from '@/fetchers/mutators';
 import { DeleteButton } from './DeleteButton/DeleteButton';
 import { EditReviewSection } from './EditReviewSection/EditReviewSection';
-import { ShowDetailsDesktop } from '../../shows/ShowDetails/layouts/ShowDetails.desktop';
-import { ShowDetailsMobile } from '../../shows/ShowDetails/layouts/ShowDetails.mobile';
-import { ReviewItemDesktop } from './layouts/ReviewItem.desktop';
-import { ReviewItemMobile } from './layouts/ReviewItem.mobile';
 
 export interface IReview {
   comment: string;
@@ -47,13 +41,91 @@ export const ReviewItem = ({ review }: IReviewItemProps) => {
     review.user.id == -1 ||
     localStorage.getItem('tv-shows-uid') == review.user?.id + '';
   return (
-    <>
-      <Show above="1024px">
-        <ReviewItemDesktop review={review} />
-      </Show>
-      <Hide above="1024px">
-        <ReviewItemMobile review={review} />
-      </Hide>
-    </>
+    <Flex
+      height={{ base: '120px', lg: '104px' }}
+      width={{ base: '344px', lg: '870px' }}
+      backgroundColor="purple2"
+      color="white"
+      borderRadius="20px"
+      padding={{ base: '24px', lg: '4 4 2 2' }}
+      flexDirection={{ base: 'column', lg: 'row' }}
+      gap={2}
+    >
+      <Flex
+        gap={{ base: '24px', lg: 2 }}
+        flexDirection={{ base: 'row', lg: 'row' }}
+        alignItems={{ base: 'none', lg: 'center' }}
+        width="100%"
+      >
+        <Image
+          borderRadius="full"
+          src={review.user.image_url}
+          alt="profilna"
+          fallbackSrc="https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_640.png"
+          width="40px"
+          height="40px"
+        />
+        <Flex
+          flexDirection="column"
+          fontSize="xs"
+          width="20%"
+          height="fit-content"
+          minWidth="fit-content"
+        >
+          <Text textStyle="note"> {review.user.email} </Text>
+          <Flex flexDirection="row" alignItems="center" gap="1">
+            <Text textStyle="note">{review.rating}/5</Text>
+            <Box width={{ base: '50%', lg: '30%' }}>
+              <StarReview
+                noOfStars={review.rating}
+                isStatic={true}
+                onChange={() => {}}
+                onHover={() => {}}
+              />
+            </Box>
+          </Flex>
+        </Flex>
+        <Text fontSize="xs" hideBelow="lg">
+          {review.comment}
+        </Text>
+
+        {isMyReview ? (
+          <Flex marginLeft="auto" direction="row" gap={3}>
+            <Menu size="xs">
+              <MenuButton
+                as={Button}
+                width="24px"
+                height="24px"
+                bg="purple2"
+                color="white"
+                _hover={{ bg: 'purple2' }}
+              >
+                ⋮
+              </MenuButton>
+              <MenuList
+                minW={0}
+                w="128px"
+                height="82px"
+                p={0}
+                borderRadius="8px"
+                overflow="hidden"
+              >
+                <MenuItem w="128px" h="41px">
+                  <EditReviewSection review={review} />
+                </MenuItem>
+                <MenuItem w="128px" h="41px">
+                  <DeleteButton review={review} />
+                </MenuItem>
+              </MenuList>
+            </Menu>
+          </Flex>
+        ) : (
+          <></>
+        )}
+      </Flex>
+      <Text fontSize="xs" hideFrom="lg">
+        {review.comment}
+      </Text>
+    </Flex>
   );
 };
