+++ tv_shows_app/src/components/features/shows/ShowReviewSection/ShowReviewSection.tsx
@@ -1,7 +1,7 @@
-import { Flex, Hide, Show, Text } from '@chakra-ui/react';
+import { Flex, Text } from '@chakra-ui/react';
 import { IReview } from '../../reviews/ReviewItem/ReviewItem';
-import { ShowReviewSectionDesktop } from './layouts/ShowReviewSection.desktop';
-import { ShowReviewSectionMobile } from './layouts/ShowReviewSection.mobile';
+import { ReviewForm } from '../../reviews/ReviewForm/ReviewForm';
+import { ReviewList } from '../../reviews/ReviewList/ReviewList';
 
 interface IShowReviewSectionProps {
   reviews: Array<IReview>;
@@ -13,13 +13,18 @@ export const ShowReviewSection = ({
   showId,
 }: IShowReviewSectionProps) => {
   return (
-    <>
-      <Show above="1024px">
-        <ShowReviewSectionDesktop reviews={reviews} showId={showId} />
-      </Show>
-      <Hide above="1024px">
-        <ShowReviewSectionMobile reviews={reviews} showId={showId} />
-      </Hide>
-    </>
+    <Flex
+      flexDirection={{ base: 'column', lg: 'row' }}
+      gap={{ base: '32px', lg: '26px' }}
+      width={{ base: '344px', lg: '1052px' }}
+    >
+      <Text color="white" width="175px" textStyle="body">
+        Reviews
+      </Text>
+      <Flex flexDirection="column" gap={{ base: '0', lg: '61px' }} width="100%">
+        <ReviewForm showId={showId} />
+        <ReviewList reviews={reviews} />
+      </Flex>
+    </Flex>
   );
 };
